requirements.txt
colorama
PyJWT>=2.1.0
requests